requirements.txt
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.18.0
PyMySQL==1.1.1
python-dotenv==1.0.1
rich==13.7.1
